feat: add minimatch support
