+++ cypress/e2e/spec-a.cy.js
@@ -1,3 +1,4 @@
+/// <reference types="cypress" />
 it('works', () => {
   cy.log('test a')
 })
