+++ cypress/e2e/sub/spec-1.cy.js
@@ -1,0 +1,4 @@
+/// <reference types="cypress" />
+it('fails in the subfolder sub', () => {
+  expect(false).to.be.true
+})
