+++ cypress/expected.json
@@ -1,5 +1,8 @@
 {
   "works": "pass",
   "fails": "fail",
-  "fails c": "fail"
+  "fails c": "fail",
+  "fails in the subfolder sub": "fail",
+  "finds the right channel by the filename end": "pass",
+  "finds the right channel by minimatch": "pass"
 }
